Anti-unification is the answer! Our algo seems to be behaving quite well.
Merging happens in two steps:

In order to return the nominator, we must check that the own-variable
mappings of the anti-unification of thedeletion context of the
nominator and the denominator are either phi(var,t) or phi(t,id). This means
that the deletion context of the nominator does not require anything
in specific, and can be applied immediatly to the denominator.
Care must be taken with that 'id', as it must be a true copy,
with contractions needing extra checkup. We must also check
for ins-ins conflicts on this step

In order to instantiate the denominator, we instead go over the
own-variable mappings in the anti-unification of the nominator
and denominator as a whole, trying to refine and specialize.
Care must be taken when refining for name clashes.